+++ CHANGES.rst
@@ -22,6 +22,7 @@ FIXED:
 
 * Include changes from ``behave v1.3.1`` (#1255, #1239)
 * issue #1028: use unittest.mock instead of mock (submitted by: pgajdos)
+* issue #990: use make_relpath_if_possible in sphinx_steps formatter to handle Windows cross-drive paths
 
 DOCUMENTATION: